requirements.txt
streamlit>=1.28.0
streamlit-chat>=0.1.1
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
Pillow>=9.0.0
